// src/components/Forms/Return/tipo-demanda/interfaces.ts
export interface DemandaProconReturn {
  pk?: string;
  usuario_atuando?: string;
  created_by?: string;
  cliente?: string;
  data_carimbo?: string;
  due_date?: string;
  arquivos?: ArquivoDemanda[];
  tipo_demanda?: string;
  status_demanda?: string;
  status?: string;
  perfil_demanda?: string;
  processo?: string;
  identificacao?: string;
  suit_id?: string;
  observacao?: {
    mensagem: string;
    criada_em: string;
    criada_por: string;
  }[];
}
// ...
export interface FormFile {
  id: string;
  info: {
    name: string;
    size: number;
    sizeFormatted: string;
    url: string; 
    unit: Unit;
    s3Key: string;
    s3Bucket: string;
  };
  status: {
    success: boolean | undefined;
    progress: number;
    message: string | undefined;
    isDeleting?: boolean;
  };
  allow: {
    delete: boolean;
    download: boolean;
    retryUpload: boolean;
    link: boolean;
  };
  dropzoneFile: File | FileRejection | undefined;
}
// ...
export type TipoJustificativaInfo = 'enviar-excecao-oito' | 'reprovar-demanda' | 
  'inativar-demanda' | 'enviar-esteira-oito' | 'enviar-auditoria-oito' | '';

export interface JustificativaInfo {
  tipo: TipoJustificativaInfo;
  title: Partial<{
    text: string;
  }>,
  justificativa: Partial<{
    isValid: boolean;
    value: string;
  }>
}